Calculated Matricies, Fixed Bug

+++ Resources/Shaders/skinning.vert
@@ -3,7 +3,7 @@
 layout(location = 0) in vec3 pos;
 layout(location = 1) in vec3 normal;
 layout(location = 2) in vec2 uv;
-layout(location = 3) in vec4 boneIndices;
+layout(location = 3) in uvec4 boneIndices;
 layout(location = 4) in vec4 boneWeights;
 
 uniform mat4 model;
@@ -22,9 +22,9 @@ out VS_OUT {
 void main()
 {
 	vec4 skinPos =  boneWeights.x * (skinInfo.boneTransform[boneIndices.x] * vec4(pos, 1.0f)) +
-				boneWeights.y * (skinInfo.boneTransform[boneIndices.y] * vec4(pos, 1.0f)) +
-			    boneWeights.z * (skinInfo.boneTransform[boneIndices.z] * vec4(pos, 1.0f)) +
-				boneWeights.w * (skinInfo.boneTransform[boneIndices.w] * vec4(pos, 1.0f));
+					boneWeights.y * (skinInfo.boneTransform[boneIndices.y] * vec4(pos, 1.0f)) +
+					boneWeights.z * (skinInfo.boneTransform[boneIndices.z] * vec4(pos, 1.0f)) +
+					boneWeights.w * (skinInfo.boneTransform[boneIndices.w] * vec4(pos, 1.0f));
 
 	frag.pos	= vec3(model * skinPos);
 	frag.normal	= normal;
